Remove IVoxelWorldSubsystem::GetWorld Add ReserveGrow Refactor IsObjectKeyNull Add GetObjectKeyName AsShared -> ToSharedRef Remove Tag from FVoxelDependency Cleanup runtime info Remove thread safety from graph execution Remove inline graphs Add ReinterpretCastRef_Unaligned Fix VOXEL_COUNTER_HELPER Add GetObjectKeyPathName MakeNullSharedRef -> SharedRef_Null Remove MakeVoxelFunctionCaller Refactor TVoxelInlineMap/Set Progress Progress Progress Progress Progress Fixes Refactor graph memory management Add FVoxelUtilities::CleanupRedirects Run FVoxelUtilities::CleanupRedirects on startup Fix FVoxelUtilities::CleanupRedirects Cleanup redirects Add DECLARE_VOXEL_THIS_TYPE Add CastStruct Add TVoxelObjectPtr Move FVoxelObjectPtr to cpp Remove FVoxelTransformRef Remove FVoxelVolumeGraphStamp::PreEditChangeOverrides/PostEditChangeOverrides VoxelObjectPtr fixes Add ForeachObjectReference Fix potentially unsafe FTransform inversions AttachToComponent -> SetupAttachment Fix instanced stamp component Fix FVoxelSerializationGuard Fix VOXEL_CUSTOM_COUNTER_HELPER Cleanup voxel stats Remove FVoxelDynamicMaterialParameter Cleanup FVoxelMaterialRef Refactor hash

+++ Source/VoxelCore/Public/VoxelMinimal/Utilities/VoxelObjectUtilities.h
@@ -146,7 +146,7 @@ namespace FVoxelUtilities
 	//////////////////////////////////////////////////////////////////////////////
 	//////////////////////////////////////////////////////////////////////////////
 
-	VOXELCORE_API uint32 HashProperty(const FProperty& Property, const void* DataPtr);
+	VOXELCORE_API uint64 HashProperty(const FProperty& Property, const void* DataPtr);
 	VOXELCORE_API void DestroyStruct_Safe(const UScriptStruct* Struct, void* StructMemory);
 	VOXELCORE_API void AddStructReferencedObjects(FReferenceCollector& Collector, const FVoxelStructView& StructView);
